Made a 'close menu' button component and added it to the hamburger menu and information modal
