+++ front-end/src/Components/CloseButton/close_icon_normal.png
