+++ front-end/src/Components/NavBar/HamburgerMenu/index.js
@@ -1,5 +1,6 @@
 import LoginButton from "../../LoginButton";
 import LogoutButton from "../../LogoutButton";
+import CloseButton from '../../CloseButton';
 import Profile from "../../Profile";
 import hamburgerIcon from './hamburger_icon.svg';
 import newPhotoIcon from './new_photo_icon.png';
@@ -26,6 +27,7 @@ const HamburgerMenu = ({ handleClick, handleChange, newImageURL, usersStars, get
                 <img src={hamburgerIcon} className="hamburger-icon" onClick={() => {setDisplayHamburgerMenu(!displayHamburgerMenu)}} alt="A hamburger menu icon, which is three horizontal lines in a column." />
             </div>
             <div className={hamburgerMenuClasses}>
+                <CloseButton handleClick={() => setDisplayHamburgerMenu(false)} />
                 <div className="hamburger-links">
                     <a>
                         <img src={newPhotoIcon} className="hamburger-link-icon" alt="A simple square illustration of mountains under a clear sky. In the bottom-right corner there is a circle containing a plus sign." />
